refactor(claim-write): split interactive write core out of edit_claims
Step 0 of the claim-write-engines plan: make the catalog/api interactive
write boundary visible to import-linter before any code moves between apps.

- Extract the generic engine into catalog/api/claim_write.py (execute_claims,
  execute_multi_entity_claims, ClaimSpec, scalar planning + validation,
  field-constraints). Imports no concrete catalog model.
- edit_claims.py keeps the domain spec-builders, importing ClaimSpec,
  ValidationErrors and raise_form_error back from claim_write.
- Add EntityClaims NamedTuple to name execute_multi_entity_claims's
  (entity, specs) argument.
- Add the "Interactive claim-write core does not import the catalog domain"
  forbid (full catalog domain), carrying two burn-down baselines:
  claim_write -> catalog.resolve (clears at resolve inversion) and
  claim_write -> catalog.exceptions (clears at StructuredValidationError move).
- Split tests along the same seam (test_claim_write.py) and document the
  contract in docs/AppBoundaries.md.

Pure code motion: lint-imports green (13/0), mypy clean, full catalog/
provenance/core suite green. No app, no migration, no behavior change.
