+++ backend/apps/catalog/api/corporate_entities.py
@@ -27,13 +27,9 @@ from ..models import (
     OperatingStatus,
 )
 from ._typing import CorporateEntityListAnnotations
+from .claim_write import execute_claims, raise_form_error, validate_scalar_fields
 from .constants import NameAliasQuery, PageParam
-from .edit_claims import (
-    execute_claims,
-    plan_alias_claims,
-    raise_form_error,
-    validate_scalar_fields,
-)
+from .edit_claims import plan_alias_claims
 from .entity_crud import register_entity_create, register_entity_delete_restore
 from .entity_list import paginated_list_response
 from .helpers import (
