+++ backend/apps/catalog/api/series.py
@@ -23,8 +23,8 @@ from apps.provenance.schemas import RichTextSchema
 
 from ..models import Credit, MachineModel, Series, Title
 from ._typing import HasTitleCount
+from .claim_write import execute_claims, plan_scalar_field_claims
 from .constants import NameQuery, PageParam
-from .edit_claims import execute_claims, plan_scalar_field_claims
 from .entity_crud import register_entity_create, register_entity_delete_restore
 from .entity_list import paginated_list_response
 from .helpers import (
