+++ backend/apps/catalog/api/titles.py
@@ -69,14 +69,14 @@ from ._title_facets import (
     ordered_titles,
 )
 from ._typing import CreditKey, FacetOptionDict, GameplayFeatureAgreement
-from .constants import DEFAULT_PAGE_SIZE
-from .edit_claims import (
+from .claim_write import (
     ClaimSpec,
     execute_claims,
-    plan_abbreviation_claims,
     plan_scalar_field_claims,
     raise_form_error,
 )
+from .constants import DEFAULT_PAGE_SIZE
+from .edit_claims import plan_abbreviation_claims
 from .entity_create import (
     assert_name_available,
     assert_public_id_available,
