+++ docs/AppBoundaries.md
@@ -66,7 +66,8 @@ Beyond the cross-app tiers, several intra-app invariants are enforced:
 - **Production code does not import test factories** — `test_factories` is test-only scaffolding and must not enter the runtime path.
 - **Catalog domain does not import the ingest system** — `ingestion/` is the bulk write path, a top-level consumer. The rest of catalog must not import it; only the management commands that drive ingest may, as the orchestration entrypoint.
 - **Ingest system does not import the catalog domain** — the forward mirror of the rule above: `ingestion/` depends only on core, provenance and citation. `forbidden_modules` enumerates the full catalog domain (not just the packages imported today) so a new coupling to any catalog package is caught, not only the burn-down baselines. The remaining catalog imports are a `# BASELINE` list severed step by step; when it reaches zero the boundary is sealed and the ingest system can move to its own app. `apps.catalog.api` is omitted — the api-leaf contract already baselines those edges.
+- **Interactive claim-write core does not import the catalog domain** — the interactive analog of the ingest rule. `catalog/api/claim_write.py` is the generic interactive write engine (validate fields, write claims in a ChangeSet, attach citations, resolve), split out of the per-relationship spec-builders that stay in `edit_claims.py`. It must not import the catalog domain, so it stays a reusable write surface and can later extract into its own app; `forbidden_modules` enumerates the full domain, with two `# BASELINE` edges remaining — `claim_write -> catalog.resolve` (the `resolve_after_mutation` call) and `claim_write -> catalog.exceptions` (`StructuredValidationError`) — cleared respectively when resolution dispatch moves to a provenance-local seam and when that error relocates to core.
 - **Patch front end does not depend on the ingest back end** — the patch system lowers YAML to an `IngestPlan` (the IR); `catalog.ingestion.patches` must not reach into `catalog.ingestion.apply`. See [DataArchitecture.md](DataArchitecture.md).
 - **Media internal layering** — `constants < models < {storage|processing|schemas|helpers} < authz < api`, a `layers` contract that also keeps `media.api` a leaf.
 
-Several contracts carry a `# BASELINE` block listing pre-existing violations that predate enforcement: upward imports in `App layer tiers` (e.g. `provenance -> catalog` deferred imports, `core -> accounts` leaks), domain-logic-stranded-under-`api/` in `The api layer is a leaf` (the `soft_delete` helpers consumed by catalog's ingestion layer) and the remaining `ingestion -> catalog` imports in `Ingest system does not import the catalog domain`. Each baseline is a burn-down list: fix the underlying import and delete the line. Do not add to one. import-linter errors on an unmatched `ignore_imports` entry, so a baseline left stale after its import is gone fails the build until deleted — that error is the burn-down's forcing function.
+Several contracts carry a `# BASELINE` block listing pre-existing violations that predate enforcement: upward imports in `App layer tiers` (e.g. `provenance -> catalog` deferred imports, `core -> accounts` leaks), domain-logic-stranded-under-`api/` in `The api layer is a leaf` (the `soft_delete` helpers consumed by catalog's ingestion layer) and the remaining `ingestion -> catalog` imports in `Ingest system does not import the catalog domain`. The `claim_write -> {catalog.resolve, catalog.exceptions}` edges in `Interactive claim-write core does not import the catalog domain` use the same `# BASELINE` mechanism but are a deliberately staged decoupling introduced with the contract, not drift that predates it. Each baseline is a burn-down list: fix the underlying import and delete the line. Do not add to one. import-linter errors on an unmatched `ignore_imports` entry, so a baseline left stale after its import is gone fails the build until deleted — that error is the burn-down's forcing function.
